Add MultiIsotopeCascade
This commit adds the cascade source and header files to the repository.
Here, the calculations for multi-component isotopes are performed and the
calculation of the total number of stages will take place.

Currently, only the product and tails concentrations for a given number of
stages can be determined.

Future versions will determine the optimum number of stages using a minimiser.