Add `CalculateNStages` in MultiIsotopeCascade
This commit allows to calculate the number of stages in the enriching and
in the stripping section for a given enrichment. It determines these numbers
by brute force, i.e., it looks for the number of stages such that the
relative difference between desired enrichment level and real enrichment
level in both tails and product gets minimised.

The type of `n_enriching` and `n_stripping` has been changed from `int` to
`double`. This is due to the fact that a decimal number of stages is used
during the calculation to effectively determine `n_enrich` and `n_stripping`,
respectively. However, after having called the function both variables
effectively contain integer numbers.